# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayMyDine Restaurant API
  description: |
    Multi-tenant restaurant ordering and management system API.
    
    **Security Warning**: This API currently has no authentication implemented.
    All endpoints are publicly accessible.
    
    **Multi-tenant**: Each restaurant has isolated data through subdomain-based routing.
  version: 1.0.0
  contact:
    name: PayMyDine Support
    email: support@paymydine.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://{tenant}.paymydine.com/api/v1
    description: Production server (tenant-specific)
security:
  - {} # No authentication currently implemented
paths:
...
  /order-status:
    get:
      summary: Get order status
      description: Retrieve status for a specific order
      operationId: getOrderStatus
      tags:
        - Orders
      parameters:
        - name: order_id
          in: query
          required: true
          description: Order ID
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Order status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OrderStatus'

    post:
      summary: Update order status
      description: Update status for a specific order
      operationId: updateOrderStatus
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                  example: 123
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, delivered, cancelled]
                  example: preparing
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order status updated successfully
...
  /table-info:
    get:
      summary: Get table information
      description: Retrieve table information by table ID
      operationId: getTableInfo
      tags:
        - Tables
      parameters:
        - name: table_id
          in: query
          required: true
          description: Table ID
          schema:
            type: string
            example: "7"
      responses:
        '200':
          description: Table information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Table'
...
components:
  schemas:
...
    OrderStatus:
      type: object
      properties:
        order_id:
          type: integer
          example: 123
        status_id:
          type: integer
          example: 1
        status_name:
          type: string
          example: "preparing"
        customer_status:
          type: integer
          example: 1
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    Table:
      type: object
      properties:
        table_id:
          type: integer
          example: 7
        table_name:
          type: string
          example: "Table 7"
        min_capacity:
          type: integer
          example: 2
        max_capacity:
          type: integer
          example: 4
        table_status:
          type: integer
          example: 1
        qr_code:
          type: string
          example: "ms10zQpxP"
        location_id:
          type: integer
          example: 1
...